requirements.txt
Django == 4.2.6
django-neomodel == 0.1.1
django-request-mapping == 0.0.12